chore: Address review comment
